add authorising condition

+++ resources/views/rezepte/show.blade.php
@@ -40,12 +40,15 @@
                 <br>
                 <h3>Zubereitung</h3>
                 <p class="beschreibung">{{$rezept->zubereitung}}</p>
+                <br>
                 <button class="normalbtn btn" onclick="window.location.href='/rezepte/{{$rezept->rID}}/edit'"
                         title="Rezept bearbeiten"><i class="material-icons btn_i">edit</i> Rezept bearbeiten
                 </button>
-                <button class="abortbtn btn" onclick="window.location.href='/rezepte/{{$rezept->rID}}/destroy'"
-                        title="Rezept löschen"><i class="fas fa-trash-alt btn_i"></i> Löschen
-                </button>
+                @if (Auth::user()->hasRole('admin'))
+                    <button class="abortbtn btn" onclick="window.location.href='/rezepte/{{$rezept->rID}}/destroy'"
+                            title="Rezept löschen"><i class="fas fa-trash-alt btn_i"></i> Rezepet löschen
+                    </button>
+                @endif
             </section>
         </div>
     </div>
